+++ .github/workflows/tests.yml
@@ -101,6 +101,8 @@ jobs:
           cd sdk/python && uv sync --python 3.10 --group test --frozen && source .venv/bin/activate && cd test/test_frontend_api && pytest -s --tb=short get_email.py test_dataset.py
           
       - name: Run http api tests against Elasticsearch
+        env:
+          ZHIPU_AI_API_KEY: ${{ secrets.ZHIPU_AI_API_KEY }}
         run: |
           export http_proxy=""; export https_proxy=""; export no_proxy=""; export HTTP_PROXY=""; export HTTPS_PROXY=""; export NO_PROXY=""
           export HOST_ADDRESS=http://host.docker.internal:9380
@@ -145,6 +147,8 @@ jobs:
           cd sdk/python && uv sync --python 3.10 --group test --frozen && source .venv/bin/activate && cd test/test_frontend_api && pytest -s --tb=short get_email.py test_dataset.py
 
       - name: Run http api tests against Infinity
+        env:
+          ZHIPU_AI_API_KEY: ${{ secrets.ZHIPU_AI_API_KEY }}
         run: |
           export http_proxy=""; export https_proxy=""; export no_proxy=""; export HTTP_PROXY=""; export HTTPS_PROXY=""; export NO_PROXY=""
           export HOST_ADDRESS=http://host.docker.internal:9380
