+++ sdk/python/test/conftest.py
@@ -18,7 +18,6 @@ import os
 
 import pytest
 import requests
-from libs.auth import RAGFlowHttpApiAuth
 
 HOST_ADDRESS = os.getenv("HOST_ADDRESS", "http://127.0.0.1:9380")
 
@@ -88,69 +87,3 @@ def get_auth():
 @pytest.fixture(scope="session")
 def get_email():
     return EMAIL
-
-
-@pytest.fixture(scope="session")
-def get_http_api_auth(get_api_key_fixture):
-    return RAGFlowHttpApiAuth(get_api_key_fixture)
-
-
-def get_my_llms(auth, name):
-    url = HOST_ADDRESS + "/v1/llm/my_llms"
-    authorization = {"Authorization": auth}
-    response = requests.get(url=url, headers=authorization)
-    res = response.json()
-    if res.get("code") != 0:
-        raise Exception(res.get("message"))
-    if name in res.get("data"):
-        return True
-    return False
-
-
-def add_models(auth):
-    url = HOST_ADDRESS + "/v1/llm/set_api_key"
-    authorization = {"Authorization": auth}
-    models_info = {
-        "ZHIPU-AI": {"llm_factory": "ZHIPU-AI", "api_key": "d06253dacd404180aa8afb096fcb6c30.KatwBIUpvCSml9sU"},
-    }
-
-    for name, model_info in models_info.items():
-        if not get_my_llms(auth, name):
-            response = requests.post(url=url, headers=authorization, json=model_info)
-            res = response.json()
-            if res.get("code") != 0:
-                raise Exception(res.get("message"))
-
-
-def get_tenant_info(auth):
-    url = HOST_ADDRESS + "/v1/user/tenant_info"
-    authorization = {"Authorization": auth}
-    response = requests.get(url=url, headers=authorization)
-    res = response.json()
-    if res.get("code") != 0:
-        raise Exception(res.get("message"))
-    return res["data"].get("tenant_id")
-
-
-@pytest.fixture(scope="session", autouse=True)
-def set_tenant_info(get_auth):
-    auth = get_auth
-    try:
-        add_models(auth)
-        tenant_id = get_tenant_info(auth)
-    except Exception as e:
-        raise Exception(e)
-    url = HOST_ADDRESS + "/v1/user/set_tenant_info"
-    authorization = {"Authorization": get_auth}
-    tenant_info = {
-        "tenant_id": tenant_id,
-        "llm_id": "glm-4-flash@ZHIPU-AI",
-        "embd_id": "BAAI/bge-large-zh-v1.5@BAAI",
-        "img2txt_id": "glm-4v@ZHIPU-AI",
-        "asr_id": "",
-        "tts_id": None,
-    }
-    response = requests.post(url=url, headers=authorization, json=tenant_info)
-    res = response.json()
-    if res.get("code") != 0:
-        raise Exception(res.get("message"))
